+++ kernel_tuner/cupy.py
@@ -1,5 +1,7 @@
-"""This module contains all CUDA specific kernel_tuner functions"""
+"""This module contains all Cupy specific kernel_tuner functions"""
 from __future__ import print_function
+
+import re
 
 import logging
 import time
@@ -9,24 +11,15 @@ from kernel_tuner.observers import BenchmarkObserver
 from kernel_tuner.nvml import nvml
 
 #embedded in try block to be able to generate documentation
-#and run tests without pycuda installed
+#and run tests without cupy installed
 try:
-    import pycuda.driver as drv
+    import cupy as cp
 except ImportError:
-    drv = None
-try:
-    from pycuda.compiler import SourceModule
-except ImportError:
-    SourceModule = None
-try:
-    from pycuda.compiler import DynamicSourceModule
-except ImportError:
-    DynamicSourceModule = None
+    cp = None
 
 
-class CudaRuntimeObserver(BenchmarkObserver):
+class CupyRuntimeObserver(BenchmarkObserver):
     """ Observer that measures time using CUDA events during benchmarking """
-
     def __init__(self, dev):
         self.dev = dev
         self.stream = dev.stream
@@ -35,19 +28,16 @@ class CudaRuntimeObserver(BenchmarkObserver):
         self.times = []
 
     def after_finish(self):
-        self.times.append(self.end.time_since(self.start))    #ms
+        self.times.append(cp.cuda.get_elapsed_time(self.start, self.end)) #ms
 
     def get_results(self):
-        results = {
-            "time": np.average(self.times),
-            "times": self.times.copy()
-        }
+        results = {"time": np.average(self.times), "times": self.times.copy()}
         self.times = []
         return results
 
 
-class CudaFunctions(object):
-    """Class that groups the CUDA functions on maintains state about the device"""
+class CupyFunctions(object):
+    """Class that groups the Cupy functions on maintains state about the device"""
 
     def __init__(self, device=0, iterations=7, compiler_options=None, observers=None):
         """instantiate CudaFunctions object used for interacting with the CUDA device
@@ -65,56 +55,46 @@ class CudaFunctions(object):
         """
         self.allocations = []
         self.texrefs = []
-        if not drv:
-            raise ImportError("Error: pycuda not installed, please install e.g. using 'pip install pycuda'.")
+        if not cp:
+            raise ImportError("Error: cupy not installed, please install e.g. using 'pip install cupy-cuda111', please check https://github.com/cupy/cupy.")
 
-        drv.init()
-        self.context = drv.Device(device).make_context()
+        #select device
+        self.dev = dev = cp.cuda.Device(device).__enter__()
 
         #inspect device properties
-        devprops = {str(k): v
-                    for (k, v) in self.context.get_device().get_attributes().items()}
-        self.max_threads = devprops['MAX_THREADS_PER_BLOCK']
-        cc = str(devprops.get('COMPUTE_CAPABILITY_MAJOR', '0')) + str(devprops.get('COMPUTE_CAPABILITY_MINOR', '0'))
-        if cc == "00":
-            cc = self.context.get_device().compute_capability()
-        self.cc = str(cc[0]) + str(cc[1])
+        self.devprops = dev.attributes
+        self.cc = dev.compute_capability
+        self.max_threads = self.devprops['MaxThreadsPerBlock']
+
         self.iterations = iterations
         self.current_module = None
-        self.func = None
         self.compiler_options = compiler_options or []
 
-        #select PyCUDA source module
-        if int(self.cc) >= 35:
-            self.source_mod = DynamicSourceModule
-        else:
-            self.source_mod = SourceModule
-        if not self.source_mod:
-            raise ImportError(
-                "Error: pycuda not correctly installed, please ensure pycuda is installed on the same CUDA installation as you're using right now")
-
         #create a stream and events
-        self.stream = drv.Stream()
-        self.start = drv.Event()
-        self.end = drv.Event()
+        self.stream = cp.cuda.Stream()
+        self.start = cp.cuda.Event()
+        self.end = cp.cuda.Event()
 
         #default dynamically allocated shared memory size, can be overwritten using smem_args
         self.smem_size = 0
 
         #setup observers
         self.observers = observers or []
-        self.observers.append(CudaRuntimeObserver(self))
+        self.observers.append(CupyRuntimeObserver(self))
         for obs in self.observers:
             obs.register_device(self)
 
         #collect environment information
         env = dict()
-        env["device_name"] = self.context.get_device().name()
-        env["cuda_version"] = ".".join([str(i) for i in drv.get_version()])
+        cupy_info = str(cp._cupyx.get_runtime_info()).split("\n")[:-1]
+        info_dict = {s.split(":")[0].strip():s.split(":")[1].strip() for s in cupy_info}
+        env["device_name"] = info_dict[f'Device {device} Name']
+
+        env["cuda_version"] = cp.cuda.runtime.driverGetVersion()
         env["compute_capability"] = self.cc
         env["iterations"] = self.iterations
         env["compiler_options"] = compiler_options
-        env["device_properties"] = devprops
+        env["device_properties"] = self.devprops
         self.env = env
         self.name = env["device_name"]
 
@@ -122,11 +102,8 @@ class CudaFunctions(object):
         return self
 
     def __exit__(self, *exc):
-        for gpu_mem in self.allocations:
-            if hasattr(gpu_mem, 'free'):    #if needed for when using mocks during testing
-                gpu_mem.free()
-        if hasattr(self, 'context'):
-            self.context.pop()
+        """destroy the device context"""
+        self.dev.__exit__()
 
     def ready_argument_list(self, arguments):
         """ready argument list to be passed to the kernel, allocates gpu mem
@@ -137,19 +114,19 @@ class CudaFunctions(object):
         :type arguments: list(numpy objects)
 
         :returns: A list of arguments that can be passed to an CUDA kernel.
-        :rtype: list( pycuda.driver.DeviceAllocation, numpy.int32, ... )
+        :rtype: list( cupy.ndarray, numpy.int32, ... )
         """
         gpu_args = []
         for arg in arguments:
             # if arg i is a numpy array copy to device
             if isinstance(arg, np.ndarray):
-                alloc = drv.mem_alloc(arg.nbytes)
+                alloc = cp.array(arg)
                 self.allocations.append(alloc)
                 gpu_args.append(alloc)
-                drv.memcpy_htod(gpu_args[-1], arg)
-            else:    # if not an array, just pass argument along
+            else: # if not a numpy array, just pass argument along
                 gpu_args.append(arg)
         return gpu_args
+
 
     def compile(self, kernel_instance):
         """call the CUDA compiler to compile the kernel, return the device function
@@ -162,29 +139,23 @@ class CudaFunctions(object):
         :type kernel_string: string
 
         :returns: An CUDA kernel that can be called directly.
-        :rtype: pycuda.driver.Function
+        :rtype: cupy.RawKernel
         """
         kernel_string = kernel_instance.kernel_string
         kernel_name = kernel_instance.name
 
-        try:
-            no_extern_c = 'extern "C"' in kernel_string
+        compiler_options = self.compiler_options
+        if not any(['--std=' in opt for opt in self.compiler_options]):
+            compiler_options = ['--std=c++11'] + self.compiler_options
 
-            compiler_options = ['-Xcompiler=-Wall']
-            if self.compiler_options:
-                compiler_options += self.compiler_options
+        options = tuple(compiler_options)
 
-            self.current_module = self.source_mod(kernel_string, options=compiler_options + ["-e", kernel_name],
-                                                  arch=('compute_' + self.cc) if self.cc != "00" else None, code=('sm_' + self.cc) if self.cc != "00" else None,
-                                                  cache_dir=False, no_extern_c=no_extern_c)
+        self.current_module = cp.RawModule(code=kernel_string, options=options,
+                                           name_expressions=[kernel_name])
 
-            self.func = self.current_module.get_function(kernel_name)
-            return self.func
-        except drv.CompileError as e:
-            if "uses too much shared data" in e.stderr:
-                raise Exception("uses too much shared data")
-            else:
-                raise e
+        self.func = self.current_module.get_function(kernel_name)
+        return self.func
+
 
     def benchmark(self, func, gpu_args, threads, grid):
         """runs the kernel and measures time repeatedly, returns average time
@@ -196,13 +167,13 @@ class CudaFunctions(object):
         this is to be robust against initialization artifacts and other exceptional
         cases.
 
-        :param func: A PyCuda kernel compiled for this specific kernel configuration
-        :type func: pycuda.driver.Function
+        :param func: A cupy kernel compiled for this specific kernel configuration
+        :type func: cupy.RawKernel
 
         :param gpu_args: A list of arguments to the kernel, order should match the
             order in the code. Allowed values are either variables in global memory
             or single values passed by value.
-        :type gpu_args: list( pycuda.driver.DeviceAllocation, numpy.int32, ...)
+        :type gpu_args: list( cupy.ndarray, numpy.int32, ...)
 
         :param threads: A tuple listing the number of threads in each dimension of
             the thread block
@@ -216,7 +187,7 @@ class CudaFunctions(object):
         :rtype: dict()
         """
         result = dict()
-        self.context.synchronize()
+        self.dev.synchronize()
         for _ in range(self.iterations):
             self.start.record(stream=self.stream)
             for obs in self.observers:
@@ -225,11 +196,10 @@ class CudaFunctions(object):
             self.end.record(stream=self.stream)
             for obs in self.observers:
                 obs.after_start()
-            while not self.end.query():
+            while not self.end.done:
                 for obs in self.observers:
                     obs.during()
-                time.sleep(1e-6) #one microsecond
-            self.end.synchronize()
+                time.sleep(1e-6)
             for obs in self.observers:
                 obs.after_finish()
 
@@ -250,14 +220,7 @@ class CudaFunctions(object):
         """
         logging.debug('copy_constant_memory_args called')
         logging.debug('current module: ' + str(self.current_module))
-        for k, v in cmem_args.items():
-            symbol = self.current_module.get_global(k)[0]
-            logging.debug('copying to symbol: ' + str(symbol))
-            logging.debug('array to be copied: ')
-            logging.debug(v.nbytes)
-            logging.debug(v.dtype)
-            logging.debug(v.flags)
-            drv.memcpy_htod(symbol, v)
+        raise NotImplementedError('CuPy backend does not yet support constant memory')
 
     def copy_shared_memory_args(self, smem_args):
         """add shared memory arguments to the kernel"""
@@ -270,68 +233,18 @@ class CudaFunctions(object):
             device texture memory. See tune_kernel().
         :type texmem_args: dict
         """
-
-        filter_mode_map = {
-            'point': drv.filter_mode.POINT,
-            'linear': drv.filter_mode.LINEAR
-        }
-        address_mode_map = {
-            'border': drv.address_mode.BORDER,
-            'clamp': drv.address_mode.CLAMP,
-            'mirror': drv.address_mode.MIRROR,
-            'wrap': drv.address_mode.WRAP
-        }
-
-        logging.debug('copy_texture_memory_args called')
-        logging.debug('current module: ' + str(self.current_module))
-        self.texrefs = []
-        for k, v in texmem_args.items():
-            tex = self.current_module.get_texref(k)
-            self.texrefs.append(tex)
-
-            logging.debug('copying to texture: ' + str(k))
-            if not isinstance(v, dict):
-                data = v
-            else:
-                data = v['array']
-            logging.debug('texture to be copied: ')
-            logging.debug(data.nbytes)
-            logging.debug(data.dtype)
-            logging.debug(data.flags)
-
-            drv.matrix_to_texref(data, tex, order="C")
-
-            if isinstance(v, dict):
-                if 'address_mode' in v and v['address_mode'] is not None:
-                    # address_mode is set per axis
-                    amode = v['address_mode']
-                    if not isinstance(amode, list):
-                        amode = [amode] * data.ndim
-                    for i, m in enumerate(amode):
-                        try:
-                            if m is not None:
-                                tex.set_address_mode(i, address_mode_map[m])
-                        except KeyError:
-                            raise ValueError('Unknown address mode: ' + m)
-                if 'filter_mode' in v and v['filter_mode'] is not None:
-                    fmode = v['filter_mode']
-                    try:
-                        tex.set_filter_mode(filter_mode_map[fmode])
-                    except KeyError:
-                        raise ValueError('Unknown filter mode: ' + fmode)
-                if 'normalized_coordinates' in v and v['normalized_coordinates']:
-                    tex.set_flags(tex.get_flags() | drv.TRSF_NORMALIZED_COORDINATES)
+        raise NotImplementedError('CuPy backend does not yet support constant memory')
 
     def run_kernel(self, func, gpu_args, threads, grid, stream=None):
         """runs the CUDA kernel passed as 'func'
 
-        :param func: A PyCuda kernel compiled for this specific kernel configuration
-        :type func: pycuda.driver.Function
+        :param func: A cupy kernel compiled for this specific kernel configuration
+        :type func: cupy.RawKernel
 
         :param gpu_args: A list of arguments to the kernel, order should match the
             order in the code. Allowed values are either variables in global memory
             or single values passed by value.
-        :type gpu_args: list( pycuda.driver.DeviceAllocation, numpy.int32, ...)
+        :type gpu_args: list( cupy.ndarray, numpy.int32, ...)
 
         :param threads: A tuple listing the number of threads in each dimension of
             the thread block
@@ -341,13 +254,13 @@ class CudaFunctions(object):
             of the grid
         :type grid: tuple(int, int)
         """
-        func(*gpu_args, block=threads, grid=grid, stream=stream, shared=self.smem_size, texrefs=self.texrefs)
+        func(grid, threads, gpu_args, stream=stream, shared_mem=self.smem_size)
 
     def memset(self, allocation, value, size):
         """set the memory in allocation to the value in value
 
         :param allocation: A GPU memory allocation unit
-        :type allocation: pycuda.driver.DeviceAllocation
+        :type allocation: cupy.ndarray
 
         :param value: The value to set the memory to
         :type value: a single 8-bit unsigned int
@@ -356,7 +269,7 @@ class CudaFunctions(object):
         :type size: int
 
         """
-        drv.memset_d8(allocation, value, size)
+        allocation[:] = value
 
     def memcpy_dtoh(self, dest, src):
         """perform a device to host memory copy
@@ -365,29 +278,27 @@ class CudaFunctions(object):
         :type dest: numpy.ndarray
 
         :param src: A GPU memory allocation unit
-        :type src: pycuda.driver.DeviceAllocation
+        :type src: cupy.ndarray
         """
-        if isinstance(src, drv.DeviceAllocation):
-            drv.memcpy_dtoh(dest, src)
+        if isinstance(dest, np.ndarray):
+            tmp = cp.asnumpy(src)
+            np.copyto(dest, tmp)
+        elif isinstance(dest, cp.ndarray):
+            cp.copyto(dest, src)
         else:
-            dest = src
+            raise ValueError("dest type not supported")
 
     def memcpy_htod(self, dest, src):
         """perform a host to device memory copy
 
         :param dest: A GPU memory allocation unit
-        :type dest: pycuda.driver.DeviceAllocation
+        :type dest: cupy.ndarray
 
         :param src: A numpy array in host memory to store the data
         :type src: numpy.ndarray
         """
-        if isinstance(dest, drv.DeviceAllocation):
-            drv.memcpy_htod(dest, src)
-        else:
-            dest = src
+        if isinstance(src, np.ndarray):
+            src = cp.asarray(src)
+        cp.copyto(dest, src)
 
-    units = {
-        'time': 'ms',
-        'power': 's,mW',
-        'energy': 'J'
-    }
+    units = {'time': 'ms'}