+++ kernel_tuner/strategies/genetic_algorithm.py
@@ -1,11 +1,11 @@
 """ A simple genetic algorithm for parameter search """
-from __future__ import print_function
 
 import random
+from collections import OrderedDict
 import numpy as np
 
 from kernel_tuner.strategies.minimize import _cost_func
-
+from kernel_tuner import util
 
 def tune(runner, kernel_options, device_options, tuning_options):
     """ Find the best performing kernel configuration in the parameter space
@@ -35,66 +35,75 @@ def tune(runner, kernel_options, device_options, tuning_options):
 
     options = tuning_options.strategy_options
     pop_size = options.get("popsize", 20)
-    generations = options.get("maxiter", 100)
+    generations = options.get("maxiter", 50)
     crossover = supported_methods[options.get("method", "uniform")]
     mutation_chance = options.get("mutation_chance", 10)
+
+    max_fevals = options.get("max_fevals", 100)
+
+    max_threads = runner.dev.max_threads
 
     tuning_options["scaling"] = False
     tune_params = tuning_options.tune_params
 
     best_time = 1e20
     all_results = []
+    unique_results = {}
 
-    population = random_population(pop_size, tune_params)
+    population = random_population(pop_size, tune_params, tuning_options, max_threads)
 
     for generation in range(generations):
-
-        # optionally enable something to remove duplicates and increase diversity,
-        # leads to longer execution times, but might improve robustness
-        # population = ensure_diversity(population, pop_size, tune_params)
-
-        if tuning_options.verbose:
-            print("Generation %d, best_time %f" % (generation, best_time))
 
         # determine fitness of population members
         weighted_population = []
         for dna in population:
             time = _cost_func(dna, kernel_options, tuning_options, runner, all_results)
             weighted_population.append((dna, time))
-        population = []
+
+        # population is sorted such that better configs have higher chance of reproducing
+        weighted_population.sort(key=lambda x: x[1])
 
         # 'best_time' is used only for printing
         if tuning_options.verbose and all_results:
             best_time = min(all_results, key=lambda x: x["time"])["time"]
 
-        # population is sorted such that better configs have higher chance of reproducing
-        weighted_population.sort(key=lambda x: x[1])
+        if tuning_options.verbose:
+            print("Generation %d, best_time %f" % (generation, best_time))
+            for dna in weighted_population:
+                print(dna)
+            diversity = len(population)
+            for dna1 in population:
+                for dna2 in population:
+                    if dna1 == dna2:
+                        diversity = diversity - 1
+            print(f"{diversity=}")
+
+        old_population = population[:]
+        population = []
+
+        unique_results.update({",".join([str(i) for i in dna]): time for dna, time in weighted_population})
+        if len(unique_results) >= max_fevals:
+            break
 
         # crossover and mutate
-        for _ in range(pop_size // 2):
+        while len(population) < pop_size:
             dna1, dna2 = weighted_choice(weighted_population, 2)
 
-            dna1, dna2 = crossover(dna1, dna2)
+            children = crossover(dna1, dna2)
 
-            population.append(mutate(dna1, tune_params, mutation_chance))
-            population.append(mutate(dna2, tune_params, mutation_chance))
+            for child in children:
+                child = mutate(child, tune_params, mutation_chance, tuning_options, max_threads)
+
+                if child not in population and util.config_valid(child, tuning_options, max_threads):
+                    population.append(child)
+
+                if len(population) >= pop_size:
+                    break
+
+        # could combine old + new generation here and do a selection
+
 
     return all_results, runner.dev.get_environment()
-
-
-def ensure_diversity(population, pop_size, tune_params):
-    """filter duplicates and replace with new random individuals"""
-    unique_set = []
-    unique_population = []
-    for i, dna in enumerate(population):
-        string_repr = " ".join([str(d) for d in dna])
-        if not string_repr in unique_set:
-            unique_population.append(dna)
-            unique_set.append(string_repr)
-    uniques = len(unique_set)
-    # could do different things here to increase population
-    # e.g. additional crossovers, or new random configs
-    return unique_population + random_population(pop_size - uniques, tune_params)
 
 
 def weighted_choice(population, n):
@@ -107,7 +116,7 @@ def weighted_choice(population, n):
         return int(random.betavariate(alpha, beta) * pop_size)
 
     def random_index_weighted(pop_size):
-        """might lead to problems: if there's only one valid configuration this method only returns that configuration"""
+        """use weights to increase probability of selection"""
         weights = [w for _, w in population]
         # invert because lower is better
         inverted_weights = [1.0 / w for w in weights]
@@ -129,16 +138,14 @@ def weighted_choice(population, n):
     return [population[ind][0] for ind in chosen]
 
 
-def random_population(pop_size, tune_params):
+def random_population(pop_size, tune_params, tuning_options, max_threads):
     """create a random population of pop_size unique members"""
     population = []
     option_space = np.prod([len(v) for v in tune_params.values()])
     assert pop_size < option_space
     while len(population) < pop_size:
-        dna = []
-        for i in range(len(tune_params)):
-            dna.append(random_val(i, tune_params))
-        if not dna in population:
+        dna = [random.choice(v) for v in tune_params.values()]
+        if not dna in population and util.config_valid(dna, tuning_options, max_threads):
             population.append(dna)
     return population
 
@@ -149,26 +156,33 @@ def random_val(index, tune_params):
     return random.choice(tune_params[key])
 
 
-def mutate(dna, tune_params, mutation_chance):
+def mutate(dna, tune_params, mutation_chance, tuning_options, max_threads):
     """Mutate DNA with 1/mutation_chance chance"""
-    dna_out = []
-    for i in range(len(dna)):
-        if int(random.random() * mutation_chance) == 1:
-            dna_out.append(random_val(i, tune_params))
-        else:
-            dna_out.append(dna[i])
-    return dna_out
+    dna_out = dna[:]
+    if int(random.random() * mutation_chance) == 0:
+        attempts = 20
+        while attempts > 0:
+            #decide which parameter to mutate
+            i = random.choice(range(len(dna)))
+            dna_out = dna[:]
+            dna_out[i] = random_val(i, tune_params)
+
+            if not dna_out == dna and util.config_valid(dna_out, tuning_options, max_threads):
+                return dna_out
+            attempts = attempts - 1
+    return dna
 
 
 def single_point_crossover(dna1, dna2):
     """crossover dna1 and dna2 at a random index"""
-    pos = int(random.random() * len(dna1))
+    #pos = 1+int(random.random() * (len(dna1)-2))
+    pos = int(random.random() * (len(dna1)))
     return (dna1[:pos] + dna2[pos:], dna2[:pos] + dna1[pos:])
 
 
 def two_point_crossover(dna1, dna2):
     """crossover dna1 and dna2 at 2 random indices"""
-    pos1, pos2 = sorted(random.sample(range(len(dna1)), 2))
+    pos1, pos2 = sorted(random.sample(range(1,len(dna1)-1), 2))
     child1 = dna1[:pos1] + dna2[pos1:pos2] + dna1[pos2:]
     child2 = dna2[:pos1] + dna1[pos1:pos2] + dna2[pos2:]
     return (child1, child2)
@@ -176,17 +190,10 @@ def two_point_crossover(dna1, dna2):
 
 def uniform_crossover(dna1, dna2):
     """randomly crossover genes between dna1 and dna2"""
-    child1 = []
-    child2 = []
-    for ind in range(len(dna1)):
-        p = random.random()
-        if p < 0.5:
-            child1.append(dna1[ind])
-            child2.append(dna2[ind])
-        else:
-            child2.append(dna1[ind])
-            child1.append(dna2[ind])
-    return (child1, child2)
+    ind = np.random.random(len(dna1)) > 0.5
+    child1 = [dna1[i] if ind[i] else dna2[i] for i in range(len(ind))]
+    child2 = [dna2[i] if ind[i] else dna1[i] for i in range(len(ind))]
+    return child1, child2
 
 
 def disruptive_uniform_crossover(dna1, dna2):