+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [0.4.0] - 2021-04-09
 ### Added
 - support for (lambda) function instead of list of strings for restrictions
 - support for (lambda) function instead of list for specifying grid divisors
@@ -14,6 +15,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - option to control measurements using observers
 - support for NVML tunable parameters
 - option to simulate auto-tuning searches from existing cache files
+- Cupy backend to support C++ templated CUDA kernels
+- support for templated CUDA kernels using PyCUDA backend
+- documentation on tunable parameter vocabulary
 
 ## [0.3.2] - 2020-11-04
 ### Added